Remove join against user_daily_visits
I discussed this with @neilisfragile and he agreed this would be a sensible
thing to do.

A small subset of users don't make it into user_daily_visists, either because
they register for the service but never manage to use it, or due to races in
calculating user_daily_visits. Here isn't the place to fix bugs in
user_daily_visits, and we ought to treat users who register for the service but
don't use it as part of the unretained cohort.

This will lead to a small decrease in retention metrics.